requirements.txt
streamlit>=1.28.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
imagehash>=4.3.0
Pillow>=10.0.0
pandas>=2.0.0
jinja2>=3.1.0
aiohttp>=3.9.0
asyncio>=3.4.3